Setup placeholder page elements
