+++ src/Pages/CompassPage.js
@@ -1,0 +1,8 @@
+
+export default function CompassPage() {
+    return (
+        <div>
+            Compass Page
+        </div>
+    )
+}
