+++ src/Pages/SelectionPage.js
@@ -1,0 +1,8 @@
+
+export default function SelectionPage() {
+    return (
+        <div>
+            Selection Page
+        </div>
+    )
+}
